Added a test crosshair and a Build Folder
Added a little circle sprite that follows the player's mouse to act a crosshair and calculate the mouse's position for other objects to use. Also compiled the game for the first time so I made a folder called 'Build Folder' to hold the .exe and other stuff needed for the game to run (make sure you set your build path to that folder!).